Payment processing feature fixed
